move printManual function to seperate module
